requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.35
databases==0.9.0
aiosqlite
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.12
argon2-cffi
